Display Elevation Profile in Segment Overlay
Adds the integration with turf.js to load the elevation profile of the selected segment and to then color the graph based on the gradient.
This only splits the segment into 10, equal sized chunks for current limitations. Ideal is to have a sort-of strava-like embed that could see the gradient at each point